+++ src/components/tooltip/tooltip-interface.ts
@@ -1,1 +1,2 @@
 export type WcsTooltipPosition = 'top' | 'bottom' | 'right' | 'left';
+export type WcsTooltipAppendTo = Element | 'parent' | ((ref: Element) => Element);