+++ stories/components/tooltip/tooltip.stories.ts
@@ -1,10 +1,13 @@
 import { Meta, Story } from '@storybook/web-components';
 import { html } from 'lit-html';
-import { WcsTooltipPosition } from '../../../src/components/tooltip/tooltip-interface';
+import { WcsTooltipAppendTo, WcsTooltipPosition } from '../../../src/components/tooltip/tooltip-interface';
 import { unsafeHTML } from 'lit-html/directives/unsafe-html';
 // @ts-ignore
 import interactiveStoryDocumentation from './interactiveStoryDocumentation.md';
+// @ts-ignore
+import fullscreenStoryDocumentation from './fullscreenStoryDocumentation.md';
 import { ifDefined } from 'lit-html/directives/if-defined';
+import { createRef, ref, Ref } from 'lit-html/directives/ref';
 
 export default {
     title: 'Components/Tooltip',
@@ -15,7 +18,7 @@ const Template: Story<Partial<{ content: string, position: WcsTooltipPosition, t
     const tooltip_unique_element_id_idx = tooltip_unique_element_id++; // use to generate a unique button id for each story (doc-only)
     return html`
     <div style="padding: 50px 0 0 200px;"><!-- div to add space for the tooltip to demonstrate the positioning property -->
-        <wcs-button shape="small" id=${`tooltiped-button-${tooltip_unique_element_id_idx}`}>Click me</wcs-button>
+        <wcs-button shape="small" id=${`tooltiped-button-${tooltip_unique_element_id_idx}`}>Hover to show tooltip</wcs-button>
         <wcs-tooltip
             for=${`tooltiped-button-${tooltip_unique_element_id_idx}`}
             max-width=${ifDefined(args.maxWidth)}
@@ -81,5 +84,48 @@ Interactive.args = {
 <a href="https://www.logoden-biniou.com/" style="display: block; color: var(--wcs-primary); margin: 12px 0" target="_blank">logoden-biniou</a>
 <wcs-button shape="small">C\'est un grand oui !</wcs-button>`
 };
+
+const fullscreenDiv: Ref<HTMLInputElement> = createRef();
+
+function enableFullScreen(e) {
+    e.currentTarget.parentNode?.requestFullscreen();
+}
+
+const FullScreenTemplate: Story<Partial<{ appendTo: WcsTooltipAppendTo, content: string, position: WcsTooltipPosition, tooltipInnerHtml: string, interactive: boolean, maxWidth: string | number, theme: string, delay: number | [number, number], duration: number | [number, number], trigger: string }>> = (args) => {
+    const tooltip_unique_element_id_idx = tooltip_unique_element_id++; // use to generate a unique button id for each story (doc-only)
+    return html`
+    <div style="padding: 50px 0 0 200px;"><!-- div to add space for the tooltip to demonstrate the positioning property -->
+        <wcs-button shape="small" id=${`tooltiped-fullscreen-button-${tooltip_unique_element_id_idx}`} @click="${enableFullScreen}">Enable Fullscreen</wcs-button>
+        <wcs-button shape="small"  ${ref(fullscreenDiv)} id=${`tooltiped-button-${tooltip_unique_element_id_idx}`}>Hover to show tooltip</wcs-button>
+        <wcs-tooltip
+            append-to=${ifDefined(args.appendTo)}
+            for=${`tooltiped-button-${tooltip_unique_element_id_idx}`}
+            max-width=${ifDefined(args.maxWidth)}
+            ?interactive=${args.interactive}
+            theme=${ifDefined(args.theme)}
+            .delay=${ifDefined(args.delay)}
+            .duration=${ifDefined(args.duration)}
+            content=${ifDefined(args.content)}
+            trigger=${ifDefined(args.trigger)}
+            position=${ifDefined(args.position)}>
+            ${unsafeHTML(args.tooltipInnerHtml)}
+        </wcs-tooltip>
+    </div>`
+};
+
+export const FullScreenAppendToParent = FullScreenTemplate.bind({});
+FullScreenAppendToParent.parameters = {
+    docs: {
+        description: {
+            story: fullscreenStoryDocumentation,
+        }
+    }
+};
+
+FullScreenAppendToParent.args = {
+    content: 'Example with a tooltip in a fullscreen Element',
+    appendTo: 'parent'
+};
+
 
 let tooltip_unique_element_id = 0;